requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
anndata==0.10.5.post1
annotated-types==0.6.0
appdirs==1.4.4
argcomplete==3.5.0
array_api_compat==1.4.1
async-timeout==4.0.3
attrs==23.2.0
cachetools==5.3.3
causal-conv1d==1.2.0.post2
certifi @ file:///croot/certifi_1707229174982/work/certifi
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
deepspeed==0.13.2
docker-pycreds==0.4.0
docstring-parser==0.15
einops==0.7.0
exceptiongroup==1.2.0
filelock @ file:///croot/filelock_1700591183607/work
flash-attn==2.5.3
fonttools==4.48.1
frozenlist==1.4.1
fsspec==2024.2.0
get-annotations==0.1.2
gitdb==4.0.11
GitPython==3.1.41
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645438755360/work
h5py==3.10.0
hjson==3.1.0
huggingface-hub==0.20.3
idna @ file:///croot/idna_1666125576474/work
importlib-resources==6.1.1
Jinja2 @ file:///croot/jinja2_1706733616596/work
joblib==1.3.2
jsonargparse==4.27.5
kaleido==0.2.1
kiwisolver==1.4.5
lightning==2.2.0.post0
lightning-utilities==0.10.1
llvmlite==0.42.0
mamba-ssm==1.2.0.post1
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib==3.8.2
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mpmath @ file:///croot/mpmath_1690848262763/work
multidict==6.0.5
munkres==1.1.4
natsort==8.4.0
networkx @ file:///croot/networkx_1690561992265/work
ninja==1.11.1.1
numba==0.59.0
numpy @ file:///croot/numpy_and_numpy_base_1704311704800/work/dist/numpy-1.26.3-cp39-cp39-linux_x86_64.whl#sha256=93e7b9e5e2090dd03810e7c1b02cb077d3ef49fc713f0af531e0667375d9decb
nvidia-ml-py==12.535.161
nvitop==1.3.2
openTSNE==1.0.2
packaging==23.2
pandas==2.2.0
patsy==0.5.6
pillow @ file:///croot/pillow_1707233021655/work
pipx==1.7.1
platformdirs==4.2.2
plotly==5.18.0
protobuf==4.25.2
psutil==5.9.8
py-cpuinfo==9.0.0
pydantic==2.6.1
pydantic_core==2.16.2
pynndescent==0.5.11
pynvml==11.5.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytorch-lightning==1.9.0
pytz==2024.1
PyYAML @ file:///croot/pyyaml_1698096049011/work
regex==2023.12.25
requests @ file:///croot/requests_1707355572290/work
safetensors==0.4.2
scanpy==1.9.8
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.13.2
sentry-sdk==1.40.3
session-info==1.0.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
statsmodels==0.14.1
stdlib-list==0.10.0
sympy @ file:///croot/sympy_1701397643339/work
tenacity==8.2.3
termcolor==2.4.0
threadpoolctl==3.2.0
tokenizers==0.15.2
tomli==2.0.2
torch==2.2.0
torchaudio==2.2.0
torchmetrics==1.3.0.post0
torchvision==0.17.0
tqdm==4.66.2
transformers==4.37.2
triton==2.2.0
typeshed-client==2.4.0
typing_extensions @ file:///croot/typing_extensions_1705599297034/work
tzdata==2024.1
umap-learn==0.5.5
urllib3 @ file:///croot/urllib3_1700840488771/work
userpath==1.9.2
wandb==0.16.0
yarl==1.9.4
zipp==3.17.0
